requirements.txt
setuptools~=47.1.1
Django~=3.0.7
djangorestframework~=3.11.0
Markdown~=3.2.2